+++ packages/lisk-transactions/src/transactions/5_dapp_transaction.ts
@@ -16,7 +16,7 @@ import * as BigNum from 'browserify-bignum';
 import { DAPP_FEE } from '../constants';
 import { TransactionError, TransactionMultiError } from '../errors';
 import { Account, Status, TransactionJSON } from '../transaction_types';
-import { createBaseTransaction, CreateBaseTransactionInput } from '../utils';
+import { CreateBaseTransactionInput } from '../utils';
 import {
 	isTypedObjectArrayWithKeys,
 	stringEndsWith,
@@ -175,34 +175,6 @@ export class DappTransaction extends BaseTransaction {
 		}
 		this.asset = tx.asset as DappAsset;
 		this._fee = new BigNum(DAPP_FEE);
-	}
-
-	public static create(input: CreateDappInput): object {
-		const { passphrase, secondPassphrase, options } = input;
-
-		const transaction = {
-			...createBaseTransaction(input),
-			type: 5,
-			fee: DAPP_FEE.toString(),
-			asset: {
-				dapp: options,
-			},
-		};
-
-		if (!passphrase) {
-			return transaction;
-		}
-
-		const transactionWithSenderInfo = {
-			...transaction,
-			senderId: transaction.senderId as string,
-			senderPublicKey: transaction.senderPublicKey as string,
-		};
-
-		const dappTransaction = new DappTransaction(transactionWithSenderInfo);
-		dappTransaction.sign(passphrase, secondPassphrase);
-
-		return dappTransaction.toJSON();
 	}
 
 	protected getAssetBytes(): Buffer {
